feat: add a basic CRUD DB for pipelines
This implements a basic Create-Read-Update-Delete database for pipeline
statuses. This can be used to check that status of a particular
pipeline, so that it could potentially be revived after a killed job.

Some caveats:
* Python ID is used for each pipeline. I am not sure if this is a good
  idea since the ID might be unique for each run.

The `create` is somewhat wrong, it could also update and is not strictly
separated. Can be fixed later on.

Tests auto-generated by GitHub, seem to cover at least the very basics.
More sophisticated tests will come later on.

Signed-off-by: Paul Gierz <pgierz@awi.de>