feat: new API for inputs in Rule
Rule has a new API. Instead of input patterns as a list of strings, you
now provide inputs, as a list of dictionaries. These should be
key/values with path and pattern as keys, later used to get the
appropriate file(s) from each path. Tests are also updated to reflect the
new way of loading files into the Rules.

BREAKING CHANGE: Any signature constructing a Rule object will need to
be changed!