Update src/pymorize/global_attributes.py
The `next(iter(thing))` is to get the first element from the thing. The simpler way to get the first element is simply index accessing `thing[0]` but over time my style of programming has moved from away from index access to iter based thing basically to avoid hardcoding 0 or 1 or 2. May be it is matter of taste but more importantly I like the way it is re-written sounds more meaningful to follow the code and I like it.

Co-authored-by: Paul Gierz <pgierz@awi.de>